Replaced text input with selection input
It was required to replace the text fields with word pieces selection.
For this reason the InputBuilder class is now being used as user input
to be compared with the data owned by WordPairOwner. To be implemented
in the next steps: modifying the user input by "erasing" wordpieces from
each word by clicking on the first piece to be erased (and then
re-adding the pieces in the required order)