Added assert for object parenting
It was required to parent all objects derived by QObject class. The only
non-parented ones would be the presenter and the ones used in unit
testing. For some of the objects parenting information is used for
specific checks (e.g. the non-parented DataSource might not be connected
to any file for testing purposes only)