Created a data entry statistics item
This item was required for having statistics reported at data entry,
like for example number of currently cached entries and number of last
saved entries. It was required that this statistic unit is separated
from manager or facade. Also it was required to have it separate from
cache as it should reside in the same thread (main) with the manager and
facade