Fixed a bug that occurred in prompt pane
The bug occurred when no valid data had been loaded. When the user
entered data, actioned the back button and responded with yes so data
was saved, the game would return to intro pane but the old no data
message was still being shown for a brief moment. This was because the return
to intro pane occured before data save was finished. The fix was
required for preventing this glitch. Also some refactoring was performed
regarding data entry so consistency is maintained