Created presenter/facade functions for language
This was required as part of the multiple language functionality
implementation task. For the moment only the "surface functionality" has
been implemented, namely changing the language into dropdown with status
updated for intro and main panes. When being changed into intro/main,
the language is also updated in data entry pane (vice-versa not valid).
When data entry pane is left it was required that the language is
restored to default (the language previously set in intro/main)