Implemented multiple languages in main/intro
This was required as part of the multiple languages feature
implementation. It is now possible to switch between multiple
languages and get word pairs for each of them (if available in the DB for that specific language) in the
game intro and main panes. It is still not possible to enter new word
pairs on a per-language basis. Until this functionality is implmented
mock data is required in the database