Solved a bug related to adding new word pairs
It was required that the intro pane gets updated with the message prompting the user to play even when it is not the active pane. Unless doing so this message would never get displayed when the user returns from data entry pane to intro pane and no unsaved data exists at return time. The bug was not present when the user exited the data entry pane without previously saving data, was prompted to save and clicked yes